requirements.txt
Django==4.2.13
celery==5.4.0
redis==5.0.4
flower==2.0.1
requests==2.31.0
python-dotenv==1.0.1
python-pptx==0.6.23
djangorestframework
django-bootstrap-v5
django-bootstrap-icons
openai